Added user location dedection on MainScreen Controller Taxi fixed some bugs
