+++ Ride-Foodv02/MainScreenMVC/FoodTaxiView.xib
@@ -11,11 +11,11 @@
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
         <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="FoodTaxiView" customModule="Ride_Foodv02" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="390" height="844"/>
+            <rect key="frame" x="0.0" y="0.0" width="390" height="140"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <stackView opaque="NO" contentMode="scaleToFill" distribution="fillEqually" spacing="20" translatesAutoresizingMaskIntoConstraints="NO" id="h3s-og-4ea">
-                    <rect key="frame" x="25" y="720" width="340" height="80"/>
+                    <rect key="frame" x="25" y="50" width="340" height="80"/>
                     <subviews>
                         <imageView clipsSubviews="YES" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Taxi" translatesAutoresizingMaskIntoConstraints="NO" id="tPN-we-tgl">
                             <rect key="frame" x="0.0" y="0.0" width="160" height="80"/>
@@ -29,7 +29,7 @@
                     </subviews>
                 </stackView>
                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" minimumScaleFactor="0.10000000000000001" translatesAutoresizingMaskIntoConstraints="NO" id="WAl-bc-hmy">
-                    <rect key="frame" x="195" y="680" width="0.0" height="20"/>
+                    <rect key="frame" x="195" y="10" width="0.0" height="20"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="20" id="Upe-5e-BMb"/>
                     </constraints>
@@ -38,7 +38,7 @@
                     <nil key="highlightedColor"/>
                 </label>
                 <imageView hidden="YES" clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Annotation" translatesAutoresizingMaskIntoConstraints="NO" id="rzY-s3-JzB">
-                    <rect key="frame" x="175" y="678" width="12" height="19"/>
+                    <rect key="frame" x="175" y="8" width="12" height="19"/>
                     <constraints>
                         <constraint firstAttribute="width" constant="12" id="1IZ-sK-ZPx"/>
                         <constraint firstAttribute="height" constant="19" id="eT8-GT-Z3z"/>
@@ -58,13 +58,14 @@
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="h3s-og-4ea" secondAttribute="bottom" constant="10" id="QCP-J6-BGr"/>
                 <constraint firstItem="h3s-og-4ea" firstAttribute="top" secondItem="rzY-s3-JzB" secondAttribute="bottom" constant="23" id="sAW-Ln-Md6"/>
             </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
             <connections>
                 <outlet property="foodImageView" destination="TQ7-P6-hZh" id="5rl-v9-HIr"/>
                 <outlet property="placeAnnotationView" destination="rzY-s3-JzB" id="HDE-Z5-fYc"/>
                 <outlet property="placeLabel" destination="WAl-bc-hmy" id="et7-zz-wMJ"/>
                 <outlet property="taxiImageView" destination="tPN-we-tgl" id="Fic-wp-eHK"/>
             </connections>
-            <point key="canvasLocation" x="131.8840579710145" y="89.732142857142847"/>
+            <point key="canvasLocation" x="45" y="128"/>
         </view>
     </objects>
     <resources>
